+++ ICD.Connect.Displays.SPlus/Devices/Simpl/AbstractSimplDisplaySettings.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Connect.Devices.Simpl;
+
+namespace ICD.Connect.Displays.SPlus.Devices.Simpl
+{
+	public abstract class AbstractSimplDisplaySettings : AbstractSimplDeviceSettings
+	{
+	}
+}