+++ ICD.Connect.Displays.SPlus/SPlusInterfaces/SPlusDisplayInterface.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Common.Properties;
+using ICD.Connect.Displays.SPlus.Devices.Simpl;
+
+namespace ICD.Connect.Displays.SPlus.SPlusInterfaces
+{
+	[PublicAPI("S+")]
+	public sealed class SPlusDisplayInterface : AbstractSPlusDisplayInterface<ISimplDisplay>
+	{
+	}
+}