+++ ICD.Connect.Displays/Devices/IDisplayWithAudio.cs
@@ -2,18 +2,19 @@
 using ICD.Common.Properties;
 using ICD.Common.Utils;
 using ICD.Common.Utils.EventArguments;
-using ICD.Connect.Devices.Controls;
 
 namespace ICD.Connect.Displays.Devices
 {
 	public interface IDisplayWithAudio : IDisplay
 	{
 		/// <summary>
-		/// Gets the volume control for this display.
+		/// Raised when the volume changes.
 		/// </summary>
-		IVolumeDeviceControl VolumeControl { get; }
-
 		event EventHandler<FloatEventArgs> OnVolumeChanged;
+
+		/// <summary>
+		/// Raised when the mute state changes.
+		/// </summary>
 		event EventHandler<BoolEventArgs> OnMuteStateChanged;
 
 		#region Properties
@@ -60,13 +61,35 @@ namespace ICD.Connect.Displays.Devices
 
 		#region Methods
 
+		/// <summary>
+		/// Sets the current volume.
+		/// </summary>
+		/// <param name="raw"></param>
 		void SetVolume(float raw);
 
+		/// <summary>
+		/// Increments the current volume.
+		/// </summary>
 		void VolumeUpIncrement();
+
+		/// <summary>
+		/// Decrements the current volume.
+		/// </summary>
 		void VolumeDownIncrement();
 
+		/// <summary>
+		/// Mutes the display.
+		/// </summary>
 		void MuteOn();
+
+		/// <summary>
+		/// Unmutes the display.
+		/// </summary>
 		void MuteOff();
+
+		/// <summary>
+		/// Toggles the current mute state on the display.
+		/// </summary>
 		void MuteToggle();
 
 		#endregion