+++ CHANGELOG.md
@@ -3,3 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+### Added
+ - Adding SPlus display interfaces
+ - Adding Simpl display interfaces
+ - Adding Simpl display device
+ - Adding Simpl display with audio device
+
+### Removed
+ - Removed old, unused SPlus display interface
+ - Removing unused VolumeControl property from IDisplayWithAudio